Move mocks to right places
Mocks moved due to requirements of mock and header files placement.
