+++ src/components/include/security_manager/security_query.h
