+++ src/components/protocol_handler/test/CMakeLists.txt
@@ -31,10 +31,8 @@
 if(BUILD_TESTS)
 
 include_directories(
+  include
   ${GMOCK_INCLUDE_DIRECTORY}
-  ${COMPONENTS_DIR}/protocol_handler/include
-  ${COMPONENTS_DIR}/protocol_handler/test/include
-  ${COMPONENTS_DIR}/include/protocol
 )
 
 set(LIBRARIES
