+++ src/components/security_manager/test/security_manager_test.cc
@@ -39,7 +39,7 @@
 
 #include "protocol_handler/mock_protocol_handler.h"
 #include "protocol_handler/mock_session_observer.h"
-#include "transport_manager_mock.h"
+#include "transport_manager/mock_transport_manager.h"
 #include "security_manager/mock_security_manager.h"
 #include "security_manager/mock_ssl_context.h"
 #include "security_manager/mock_crypto_manager.h"
