+++ src/components/transport_manager/test/tcp_client_listener_test.cc
@@ -33,7 +33,7 @@
 #include "gtest/gtest.h"
 #include "include/transport_adapter_mock.h"
 #include "transport_manager/tcp/tcp_client_listener.h"
-#include "include/transport_manager_mock.h"
+#include "transport_manager/mock_transport_manager.h"
 #include "transport_manager/transport_adapter/transport_adapter_controller.h"
 #include "transport_manager/transport_adapter/device.h"
