+++ src/components/transport_manager/test/transport_adapter_listener_test.cc
@@ -33,7 +33,7 @@
 #include "gtest/gtest.h"
 #include "include/transport_adapter_mock.h"
 #include "transport_manager/transport_adapter/transport_adapter_listener_impl.h"
-#include "include/transport_manager_mock.h"
+#include "transport_manager/mock_transport_manager.h"
 
 namespace test {
 namespace components {
@@ -52,7 +52,7 @@ class TransportAdapterListenerTest : public ::testing::Test {
  protected:
   const int app_handle;
   const std::string dev_id;
-  TransportManagerMock tr_mock;
+  MockTransportManager tr_mock;
   TransportAdapterMock adapter_mock;
   TransportAdapterListenerImpl transport_listener;
 };
